+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.1.0
+
+* Added: component.clone()
+* Added: component.options.cloned()
+* Removed: component.extend()
+
+# 1.0.1 
+
+* Fixed: Bug treating isMounted as a method instead of a property.
+
 # 1.0.0 (7/16/2017)
 
 * Initial public release.